remove: DogProfileComponent 하나로 줄이기(정리)  (#5)
