+++ src/app/(main)/home/mbti/page.jsx
@@ -1,19 +1,48 @@
-import DogProfileApi from '../../../../api/DogProfileApi/DogProfileApi';
+'use client'
+
+import styled from 'styled-components';
+
+import CustomButton from '../../../../components/common/CustomButton';
 import InfoBanner from '../../../../components/common/InfoBanner';
+import SectionDivider from '../../../../components/common/SectionDivider';
 import CustomHeader from '../../../../components/layout/CustomHeader';
+import { BUTTON_COLORS } from '../../../../constants/colors';
+/*import { FONTS } from '../../../../constants/font';*/
+import { PADDING_HORIZONTAL, PADDING_VERTICAL } from '../../../../constants/space';
+import DogProfileComponent from '../DogProfile';
+import MbtiMatchComponent from './MbtiMatch';
 
 const MbtiMatch = () => {
     return (
         <>
             <CustomHeader title="MBTI 매칭" showBackBtn={true} />
             <InfoBanner message="나와 멍BTI가 맞는 찰떡궁합 멍즈를 찾았어요! 🐶" />
-            <DogProfileApi subtitle = "나와 맞는 멍즈를 찾았어요! 🐾"/>
-            <div>
-                <h2>홈</h2>
-                <p>mbti 매칭</p>
-            </div>
+            <MbtiMatchComponent />
+            <SectionDivider/>
+            <DogProfileComponent subtitle = "나와 맞는 멍즈를 찾았어요! 🐾"/>
+            <ContentContainer>
+                <ButtonWrapper>
+                    <CustomButton color={BUTTON_COLORS.primary} text="동행 신청하기" route="/home/info" />
+                    <CustomButton color={BUTTON_COLORS.secondary} text="보호소로 문의하기" route="/home/mbti" />
+                </ButtonWrapper>
+            </ContentContainer>
         </>
     );
 };
 
 export default MbtiMatch;
+
+
+const ContentContainer = styled.div`
+    ${PADDING_HORIZONTAL}
+    ${PADDING_VERTICAL}
+    display: flex;
+    flex-direction: column;
+    gap: 0.63rem;
+`;
+
+const ButtonWrapper = styled.div`
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+`;
